+++ feature/alarmsetting/src/main/java/com/inseoul/alarmsetting/AlarmSettingViewModel.kt
@@ -10,18 +10,13 @@ class AlarmSettingViewModel : ViewModel() {
     val alarmDay: List<String> = _alarmDay
 
     fun addAlarmDay(day: String) {
-        // _alarmDay.add(day)
-
         if (_alarmDay.contains(day)) {
             _alarmDay.remove(day)
         } else {
             _alarmDay.add(day)
         }
         if (_alarmDay.isNullOrEmpty()) {
-            Log.d("TAG", "addAlarmDay: 결과를 입력해주세요")
             _alarmDay.add("")
         }
-
-        Log.d("AlarmSettingVM", "addAlarmDay: ${_alarmDay.last()}")
     }
 }
