+++ feature/stretching/src/main/java/com/example/stretching/timer/Timer.kt
@@ -16,6 +16,10 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.LottieConstants
+import com.airbnb.lottie.compose.animateLottieCompositionAsState
+import com.airbnb.lottie.compose.rememberLottieComposition
 import com.example.common.Constants
 import com.example.stretching.timer.TimerViewModel
 import com.inseoul.onetwothree.ui.theme.Typography
@@ -56,38 +60,7 @@ fun StretchingTimer(time: String, remainingTime: Long) {
                 style = Typography.h1
             )
         }
-        Text(text = "시간에 따른 텍스트 변화 테스트")
-        if (changeTimeFirst(remainingTime)) {
-            Text(text = "1 번")
-        } else if (changeTimeSecond(remainingTime)) {
-            Text(text = "2번")
-        } else if (changeTimeThird(remainingTime)) {
-            Text(text = "3번")
-        }
-        
-
-
     }
-    if (isTimeFinish(remainingTime)) {
-        // 화면 이동
-    }
-    /*Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = time,
-            fontSize = 60.sp,
-            color = Color.Black,
-            style = Typography.h1
-        )
-    }
-    if (isTimeFinish(remainingTime)) {
-        // 화면 이동
-    }*/
 }
 
 @RequiresApi(Build.VERSION_CODES.O)
@@ -134,6 +107,23 @@ fun TimerButton(timerState: TimerViewModel) {
     }
 }
 
+@Composable
+fun LottieAnimation(res: Int) {
+    val composition by rememberLottieComposition(
+        spec = LottieCompositionSpec.RawRes(
+            res
+        ),
+    )
+    val progress by animateLottieCompositionAsState(
+        composition = composition,
+        iterations = LottieConstants.IterateForever
+    )
+    com.airbnb.lottie.compose.LottieAnimation(
+        composition = composition,
+        progress = { progress }
+    )
+}
+
 @RequiresApi(Build.VERSION_CODES.O)
 fun Duration.format(): String {
     val seconds = kotlin.math.abs(seconds)
@@ -147,7 +137,7 @@ fun Duration.format(): String {
 
 fun isTimeFinish(time: Long) = time < 1000L
 fun changeTimeFirst(time: Long) = time in 20001..24999
-fun changeTimeSecond(time: Long) = time < 20000L && 15000L<time
+fun changeTimeSecond(time: Long) = time in 15001..19999
 fun changeTimeThird(time: Long) = time < 15000L
 
 @RequiresApi(Build.VERSION_CODES.O)
